requirements.txt
asgiref==3.3.1
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Deprecated==1.2.12
dj-database-url==0.5.0
Django==3.1.7
django-appconf==1.0.4
django-googledrive-storage==1.5.2
django-heroku==0.3.1
django-otp==1.0.2
django-phone-verify==2.0.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
firebase-admin==4.5.3
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.2
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
msgpack==1.0.2
nexmo==2.5.2
oauth2client==4.1.3
packaging==20.9
phonenumbers==8.12.19
Pillow==8.1.2
proto-plus==1.18.1
protobuf==3.15.6
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
twilio==6.53.0
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
wrapt==1.12.1
